+++ Onboard-SDK/project_folder/transceiver_search/transceiver_search.hpp
@@ -47,17 +47,18 @@
 bool setUpSubscription(DJI::OSDK::Vehicle* vehicle, int responseTimeout);
 bool teardownSubscription(DJI::OSDK::Vehicle* vehicle, const int pkgIndex, int responseTimeout);
 
-bool runWaypointMission(DJI::OSDK::Vehicle* vehicle, uint8_t numWaypoints, int responseTimeout, int S, int W);
+bool runWaypointMission(DJI::OSDK::Vehicle* vehicle, uint8_t numWaypoints, int responseTimeout, float S, float W);
 
 void setWaypointDefaults(DJI::OSDK::WayPointSettings* wp);
 void setWaypointInitDefaults(DJI::OSDK::WayPointInitSettings* fdata);
 
-std::vector<DJI::OSDK::WayPointSettings> createWaypoints(DJI::OSDK::Vehicle* vehicle, int numWaypoints, int S, int W,
-                                                         DJI::OSDK::float64_t distanceIncrement,
+std::vector<DJI::OSDK::WayPointSettings> createWaypoints(DJI::OSDK::Vehicle* vehicle, int numWaypoints, float S,
+                                                         float W, DJI::OSDK::float64_t distanceIncrement,
                                                          DJI::OSDK::float32_t start_alt);
 
 std::vector<DJI::OSDK::WayPointSettings> generateWaypoints(DJI::OSDK::WayPointSettings* start_data,
-                                                           DJI::OSDK::float64_t increment, int num_wp, int S, int W);
+                                                           DJI::OSDK::float64_t increment, int num_wp, float S,
+                                                           float W);
 
 void uploadWaypoints(DJI::OSDK::Vehicle* vehicle, std::vector<DJI::OSDK::WayPointSettings>& wp_list,
                      int responseTimeout);
